phase four completed
moved the nasa api code to a separate component

then moved the component to a separate file
